Fixed up about section mobile stylings
