+++ src/components/about/About.tsx
@@ -1,13 +1,10 @@
 import * as React from 'react';
-import Button from '@material-ui/core/Button';
 import ScrollableSection from 'react-update-url-on-scroll';
 import { graphql, StaticQuery } from 'gatsby';
 import {
   AboutContainer,
   ParagraphsContainer,
   Portrait,
-  ResumeButtonStyles,
-  StyledAnchor,
   StyledHeader,
   StyledParagraph
 } from './About.style';
@@ -23,10 +20,10 @@ export default class About extends React.Component {
           <Portrait fluid={data.about.childImageSharp.fluid} alt='about'/>
           <ParagraphsContainer>
             <StyledHeader>
-              Aloha! 
+              Aloha! Call Me Chris.
             </StyledHeader>
             <StyledParagraph>                
-              My name is Christopher Akatsuka. I grew up in Hawaii and graduated from the University of Pennsylvania 
+              I grew up in Hawaii and graduated from the University of Pennsylvania 
               with a BSE and MSE in Computer and Information Science. I am currently a Senior Associate 
               Software Engineer at Capital One in Plano, Texas where I work on full stack projects.
             </StyledParagraph>
@@ -38,11 +35,6 @@ export default class About extends React.Component {
               In my free time, I enjoy powerlifting and collecting Funko pops. I recently lifted a meet PR for 1001lbs and 
               surpassed 250 Funko pops in my collection.
             </StyledParagraph>
-            <StyledAnchor href="https://github.com/Akatsukac/akatsukac/raw/master/src/assets/docs/Final-Akatsuka-Resume.pdf" download>
-              <Button variant="outlined" size="large" style={ResumeButtonStyles}>
-                DOWNLOAD RESUME
-              </Button>
-            </StyledAnchor>
           </ParagraphsContainer>
         </AboutContainer>
       </ScrollableSection>
