+++ src/components/home/Home.tsx
@@ -1,7 +1,10 @@
 import * as React from 'react';
+import Button from '@material-ui/core/Button';
 import ScrollableSection from 'react-update-url-on-scroll';
 import { graphql, StaticQuery } from 'gatsby';
 import { 
+  ResumeButtonStyles,
+  StyledAnchor,
   StyledBackgroundImage,
   StyledHeader,
   StyledSubheader,
@@ -20,6 +23,11 @@ export default class Home extends React.Component {
             <WelcomeContainer>
               <StyledHeader>CHRISTOPHER AKATSUKA</StyledHeader>
               <StyledSubheader>Full Stack Developer</StyledSubheader>
+              <StyledAnchor href="https://github.com/Akatsukac/akatsukac/raw/master/src/assets/docs/Final-Akatsuka-Resume.pdf" download>
+                <Button variant="outlined" size="large" style={ResumeButtonStyles}>
+                  DOWNLOAD RESUME
+                </Button>
+            </StyledAnchor>
             </WelcomeContainer>
           </StyledBackgroundImage>
         </ScrollableSection>
